requirements.txt
numpy
pandas
matplotlib
seaborn
streamlit
torch
opencv-python
Pillow
ultralytics